+++ src/decode/stb_vorbis.c
@@ -37,9 +37,9 @@
 /*************************************************************************/
 
 extern stb_vorbis * stb_vorbis_open_callbacks(
-    long (*read_callback)(void *opaque, void *buf, long len),
-    void (*seek_callback)(void *opaque, long offset),
-    long (*tell_callback)(void *opaque),
+    int32_t (*read_callback)(void *opaque, void *buf, int32_t len),
+    void (*seek_callback)(void *opaque, int64_t offset),
+    int64_t (*tell_callback)(void *opaque),
     void *opaque, int64_t length, int *error_ret)
 {
     stb_vorbis *handle = mem_alloc(opaque, sizeof(*handle));
